fix[module]: rename module Cart to CustomerCart
