+++ src/Models/Mappings/MappingProfile.cs
@@ -1,7 +1,7 @@
 using AutoMapper;
-using ECommerce.Modules.Cart;
 using ECommerce.Modules.Category;
 using ECommerce.Modules.Customer;
+using ECommerce.Modules.CustomerCart;
 using ECommerce.Modules.CustomerAddress;
 using ECommerce.Modules.Product;
 using ECommerce.Modules.Sales;
