+++ src/Modules/CustomerCart/CartRepository.cs
@@ -1,4 +1,4 @@
-namespace ECommerce.Modules.Cart;
+namespace ECommerce.Modules.CustomerCart;
 
 using AutoMapper;
 using ECommerce.Context;
