+++ src/Modules/CustomerCart/CartService.cs
@@ -1,4 +1,4 @@
-namespace ECommerce.Modules.Cart;
+namespace ECommerce.Modules.CustomerCart;
 
 using ECommerce.Exceptions;
 using ECommerce.Modules.Product;
