+++ src/Modules/CustomerCart/Contracts/ICartRepository.cs
@@ -1,4 +1,4 @@
-namespace ECommerce.Modules.Cart;
+namespace ECommerce.Modules.CustomerCart;
 
 public interface ICartRepository
 {
