+++ src/Modules/CustomerCart/Contracts/ICartService.cs
@@ -1,4 +1,4 @@
-namespace ECommerce.Modules.Cart;
+namespace ECommerce.Modules.CustomerCart;
 
 public interface ICartService
 {
