+++ src/Modules/CustomerCart/DTOs/CartDTO.cs
@@ -1,4 +1,4 @@
-namespace ECommerce.Modules.Cart;
+namespace ECommerce.Modules.CustomerCart;
 
 using ECommerce.Models;
